Make layer private (#413)
* divide templates tests into parameter and circuit integration

* Add circuit integration tests for layers.

* Add circuit integration tests for embeddings.

* started interface integration tests

* avoid iteration in template

* fix device mistake

* added more interface integration tests

* automated standard tests

* automated tests

* Add qnode integration tests

* make integration tests fully automatic

* fix dict unpacking

* Found bug - integration test all running now.

* Update tests/test_templates.py


* Update tests/test_templates.py


* Update tests/test_templates.py


* Update tests/test_templates.py


* Update tests/test_templates.py


* clean up merge conflicts accidentally committed

* backup commit

* Parametrize circuits and interfaces, test each combination of templates, remove merge conflict traces

* use OrderedDict to conserve order of arguments

* move interferometer test to new file

* move interferometer test to new file

* removed reference to single layers

* remove initializations for single layer templates

* added change to CHANGELOG

* Added Maria to changelog

* backup commit

* fix python3.5 **kwargs ordering issue

* update changelog

* move changelog entry

* update changelog

* update init tests

* add numpy and torch grad tests

* add tf and cv gradient tests

* add comments to confirm that tests do not have assert statements

* remove statepreparations again from tests, will address in other PR

* make random layer also act on 1 wire

* update changelog that two templates now work with 1 wire

* address Josh's post-approval comments
