+++ tests/test_templates.py
@@ -351,7 +351,6 @@ class TestInitializationIntegration:
     @pytest.mark.parametrize("template, inpts", cv_func)
     def test_integration_cv_init(self, template, inpts, gaussian_device, n_subsystems, n_layers):
         """Checks parameter initialization compatible with continuous-variable templates."""
-
         inp = inpts(n_layers=n_layers, n_wires=n_subsystems)
         @qml.qnode(gaussian_device)
         def circuit(inp_):
@@ -409,7 +408,7 @@ class TestGradientIntegration:
         # Check gradients in numpy interface
         if intrfc == 'numpy':
             grd = qml.grad(circuit, argnum=argnm)
-            grd(*inpts)
+            assert grd(*inpts) is not None
 
         # Check gradients in torch interface
         if intrfc == 'torch':
@@ -418,11 +417,12 @@ class TestGradientIntegration:
             res = circuit(*inpts)
             res.backward()
             for a in argnm:
-                inpts[a].grad.numpy()
+                assert inpts[a].grad.numpy() is not None
 
         # Check gradients in tf interface
         if intrfc == 'tf':
             grad_inpts = [inpts[a] for a in argnm]
             with tf.GradientTape() as tape:
                 loss = circuit(*inpts)
-                tape.gradient(loss, grad_inpts)
+                assert tape.gradient(loss, grad_inpts) is not None
+
