+++ tests/test_templates_layers.py
@@ -20,313 +20,11 @@ import pytest
 import logging as log
 import pennylane as qml
 from pennylane import numpy as np
-from pennylane.qnode import QuantumFunctionError
-from pennylane.templates.layers import (Interferometer,
-                                        CVNeuralNetLayers,
-                                        StronglyEntanglingLayers, StronglyEntanglingLayer,
-                                        RandomLayers, RandomLayer)
+from pennylane.templates.layers import (CVNeuralNetLayers,
+                                        StronglyEntanglingLayers, _strongly_entangling_layer,
+                                        RandomLayers, _random_layer)
 from pennylane import RX, RY, RZ, CZ, CNOT
 log.getLogger('defaults')
-
-
-class TestInterferometer:
-    """Tests for the Interferometer from the pennylane.template.layers module."""
-
-    def test_exceptions(self):
-        """test that exceptions are correctly raised"""
-        dev = qml.device('default.gaussian', wires=1)
-        varphi = [0.42342]
-
-        @qml.qnode(dev)
-        def circuit(varphi, mesh):
-            Interferometer(theta=None, phi=None, varphi=varphi, mesh=mesh, wires=0)
-            return qml.expval(qml.NumberOperator(0))
-
-        with pytest.raises(QuantumFunctionError) as excinfo:
-            circuit(varphi, 'rectangular')
-        assert excinfo.value.args[0] == 'The mesh parameter influences the circuit architecture ' \
-                                        'and can not be passed as a QNode parameter.'
-
-        @qml.qnode(dev)
-        def circuit(varphi, bs):
-            Interferometer(theta=None, phi=None, varphi=varphi, beamsplitter=bs, wires=0)
-            return qml.expval(qml.NumberOperator(0))
-
-        with pytest.raises(QuantumFunctionError) as excinfo:
-            circuit(varphi, 'clements')
-        assert excinfo.value.args[0] == "The beamsplitter parameter influences the circuit architecture " \
-                                        "and can not be passed as a QNode parameter."
-
-    def test_clements_beamsplitter_convention(self, tol):
-        """test the beamsplitter convention"""
-        N = 2
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321]
-        phi = [0.234]
-        varphi = [0.42342]
-
-        def circuit_rect(varphi):
-            Interferometer(theta, phi, varphi, mesh='rectangular', beamsplitter='clements', wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        def circuit_tria(varphi):
-            Interferometer(theta, phi, varphi, mesh='triangular', beamsplitter='clements', wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        for c in [circuit_rect, circuit_tria]:
-            qnode = qml.QNode(c, dev)
-            assert np.allclose(qnode(varphi), [0, 0], atol=tol)
-
-            queue = qnode.queue
-            assert len(queue) == 3
-
-            assert isinstance(qnode.queue[0], qml.Rotation)
-            assert qnode.queue[0].parameters == phi
-
-            assert isinstance(qnode.queue[1], qml.Beamsplitter)
-            assert qnode.queue[1].parameters == [theta[0], 0]
-
-            assert isinstance(qnode.queue[2], qml.Rotation)
-            assert qnode.queue[2].parameters == varphi
-
-    def test_one_mode(self, tol):
-        """Test that a one mode interferometer correctly gives a rotation gate"""
-        dev = qml.device('default.gaussian', wires=1)
-        varphi = [0.42342]
-
-        def circuit(varphi):
-            Interferometer(theta=None, phi=None, varphi=varphi, wires=0)
-            return qml.expval(qml.NumberOperator(0))
-
-        qnode = qml.QNode(circuit, dev)
-        assert np.allclose(qnode(varphi), 0, atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 1
-        assert isinstance(qnode.queue[0], qml.Rotation)
-        assert np.allclose(qnode.queue[0].parameters, varphi, atol=tol)
-
-    def test_two_mode_rect(self, tol):
-        """Test that a two mode interferometer using the rectangular mesh
-        correctly gives a beamsplitter+rotation gate"""
-        N = 2
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321]
-        phi = [0.234]
-        varphi = [0.42342]
-
-        def circuit(varphi):
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        qnode = qml.QNode(circuit, dev)
-        assert np.allclose(qnode(varphi), [0, 0], atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 2
-
-        assert isinstance(qnode.queue[0], qml.Beamsplitter)
-        assert qnode.queue[0].parameters == theta+phi
-
-        assert isinstance(qnode.queue[1], qml.Rotation)
-        assert qnode.queue[1].parameters == varphi
-
-    def test_two_mode_triangular(self, tol):
-        """Test that a two mode interferometer using the triangular mesh
-        correctly gives a beamsplitter+rotation gate"""
-        N = 2
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321]
-        phi = [0.234]
-        varphi = [0.42342]
-
-        def circuit(varphi):
-            Interferometer(theta, phi, varphi, mesh='triangular', wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        qnode = qml.QNode(circuit, dev)
-        assert np.allclose(qnode(varphi), [0, 0], atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 2
-
-        assert isinstance(qnode.queue[0], qml.Beamsplitter)
-        assert qnode.queue[0].parameters == theta+phi
-
-        assert isinstance(qnode.queue[1], qml.Rotation)
-        assert qnode.queue[1].parameters == varphi
-
-    def test_two_mode_rect_overparameterised(self, tol):
-        """Test that a two mode interferometer using the rectangular mesh
-        correctly gives a beamsplitter+2 rotation gates"""
-        N = 2
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321]
-        phi = [0.234]
-        varphi = [0.42342, 0.543]
-
-        def circuit(varphi):
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        qnode = qml.QNode(circuit, dev)
-        assert np.allclose(qnode(varphi), [0, 0], atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 3
-
-        assert isinstance(qnode.queue[0], qml.Beamsplitter)
-        assert qnode.queue[0].parameters == theta+phi
-
-        assert isinstance(qnode.queue[1], qml.Rotation)
-        assert qnode.queue[1].parameters == [varphi[0]]
-
-        assert isinstance(qnode.queue[2], qml.Rotation)
-        assert qnode.queue[2].parameters == [varphi[1]]
-
-    def test_three_mode(self, tol):
-        """Test that a three mode interferometer using either mesh gives the correct gates"""
-        N = 3
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321, 0.4523, 0.21321]
-        phi = [0.234, 0.324, 0.234]
-        varphi = [0.42342, 0.234]
-
-        def circuit_rect(varphi):
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        def circuit_tria(varphi):
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        for c in [circuit_rect, circuit_tria]:
-            # test both meshes (both give identical results for the 3 mode case).
-            qnode = qml.QNode(c, dev)
-            assert np.allclose(qnode(varphi), [0]*N, atol=tol)
-
-            queue = qnode.queue
-            assert len(queue) == 5
-
-            expected_bs_wires = [[0, 1], [1, 2], [0, 1]]
-
-            for idx, op in enumerate(qnode.queue[:3]):
-                assert isinstance(op, qml.Beamsplitter)
-                assert op.parameters == [theta[idx], phi[idx]]
-                assert op.wires == expected_bs_wires[idx]
-
-            for idx, op in enumerate(qnode.queue[3:]):
-                assert isinstance(op, qml.Rotation)
-                assert op.parameters == [varphi[idx]]
-                assert op.wires == [idx]
-
-    def test_four_mode_rect(self, tol):
-        """Test that a 4 mode interferometer using rectangular mesh gives the correct gates"""
-        N = 4
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321, 0.4523, 0.21321, 0.123, 0.5234, 1.23]
-        phi = [0.234, 0.324, 0.234, 1.453, 1.42341, -0.534]
-        varphi = [0.42342, 0.234, 0.4523]
-
-        def circuit_rect(varphi):
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        qnode = qml.QNode(circuit_rect, dev)
-        assert np.allclose(qnode(varphi), [0]*N, atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 9
-
-        expected_bs_wires = [[0, 1], [2, 3], [1, 2], [0, 1], [2, 3], [1, 2]]
-
-        for idx, op in enumerate(qnode.queue[:6]):
-            assert isinstance(op, qml.Beamsplitter)
-            assert op.parameters == [theta[idx], phi[idx]]
-            assert op.wires == expected_bs_wires[idx]
-
-        for idx, op in enumerate(qnode.queue[6:]):
-            assert isinstance(op, qml.Rotation)
-            assert op.parameters == [varphi[idx]]
-            assert op.wires == [idx]
-
-    def test_four_mode_triangular(self, tol):
-        """Test that a 4 mode interferometer using triangular mesh gives the correct gates"""
-        N = 4
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        theta = [0.321, 0.4523, 0.21321, 0.123, 0.5234, 1.23]
-        phi = [0.234, 0.324, 0.234, 1.453, 1.42341, -0.534]
-        varphi = [0.42342, 0.234, 0.4523]
-
-        def circuit_tria(varphi):
-            Interferometer(theta, phi, varphi, mesh='triangular', wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        qnode = qml.QNode(circuit_tria, dev)
-        assert np.allclose(qnode(varphi), [0]*N, atol=tol)
-
-        queue = qnode.queue
-        assert len(queue) == 9
-
-        expected_bs_wires = [[2, 3], [1, 2], [0, 1], [2, 3], [1, 2], [2, 3]]
-
-        for idx, op in enumerate(qnode.queue[:6]):
-            assert isinstance(op, qml.Beamsplitter)
-            assert op.parameters == [theta[idx], phi[idx]]
-            assert op.wires == expected_bs_wires[idx]
-
-        for idx, op in enumerate(qnode.queue[6:]):
-            assert isinstance(op, qml.Rotation)
-            assert op.parameters == [varphi[idx]]
-            assert op.wires == [idx]
-
-    def test_integration(self, tol):
-        """test integration with PennyLane and gradient calculations"""
-        N = 4
-        wires = range(N)
-        dev = qml.device('default.gaussian', wires=N)
-
-        sq = np.array([[0.8734294, 0.96854066],
-                       [0.86919454, 0.53085569],
-                       [0.23272833, 0.0113988 ],
-                       [0.43046882, 0.40235136]])
-
-        theta = np.array([3.28406182, 3.0058243, 3.48940764, 3.41419504, 4.7808479, 4.47598146])
-        phi = np.array([3.89357744, 2.67721355, 1.81631197, 6.11891294, 2.09716418, 1.37476761])
-        varphi = np.array([0.4134863, 6.17555778, 0.80334114, 2.02400747])
-
-        @qml.qnode(dev)
-        def circuit(theta, phi, varphi):
-            for w in wires:
-                qml.Squeezing(sq[w][0], sq[w][1], wires=w)
-
-            Interferometer(theta, phi, varphi, wires=wires)
-            return [qml.expval(qml.NumberOperator(w)) for w in wires]
-
-        res = circuit(theta, phi, varphi)
-        expected = np.array([0.96852694, 0.23878521, 0.82310606, 0.16547786])
-        assert np.allclose(res, expected, atol=tol)
-
-        res = qml.jacobian(circuit, 0)(theta, phi, varphi)
-        expected = np.array([[-6.18547248e-03, -3.20488426e-04, -4.20274087e-02, -6.21819638e-02, 9.68526932e-01, 9.68526932e-01],
-                             [ 3.55439246e-04,  3.89820238e-02, -3.35281306e-03,  7.93009027e-04, 8.30347888e-02,-3.45150707e-01],
-                             [ 5.44893380e-03,  9.30878007e-03, -5.33374094e-01,  6.13889548e-02, -1.16931385e-01, 3.45150707e-01],
-                             [ 3.81099442e-04, -4.79703154e-02,  5.78754316e-01,  1.65477867e-01, 3.38965967e-02, 1.65477867e-01]])
-        assert np.allclose(res, expected, atol=tol)
 
 
 class TestCVNeuralNet:
@@ -507,7 +205,7 @@ class TestStronglyEntangling:
 
 
 class TestRandomLayers:
-    """Tests for the RandomLayers method from the pennylane.templates.layers module."""
+    """Tests for the RandomLayers method from the pennylane.templates module."""
 
     @pytest.fixture(scope="class",
                     params=[0.2, 0.6])
@@ -524,8 +222,8 @@ class TestRandomLayers:
     def rots(self, request):
         return request.param
 
-    def test_random_layers_seed_deterministic(self, n_layers, tol, seed):
-        """Test that pennylane.templates.layers.RandomLayers() acts deterministically when using fixed seed."""
+    def test_random_layers_deterministic_seed(self, n_layers, tol, seed):
+        """Test that RandomLayers() acts deterministically when using fixed seed."""
         n_rots = 1
         n_wires = 2
         dev = qml.device('default.qubit', wires=n_wires)
@@ -541,11 +239,10 @@ class TestRandomLayers:
 
         qnode1 = qml.QNode(circuit1, dev)
         qnode2 = qml.QNode(circuit2, dev)
-
         assert np.allclose(qnode1(weights), qnode2(weights), atol=tol)
 
-    def test_random_layers_no_seed_deterministic(self, n_layers, tol):
-        """Test that pennylane.templates.layers.RandomLayers() acts deterministically when using no seed."""
+    def test_random_layers_deterministic_default_seed(self, n_layers, tol):
+        """Test that RandomLayers() acts deterministically when using default seed."""
         n_rots = 1
         n_wires = 2
         dev = qml.device('default.qubit', wires=n_wires)
@@ -565,7 +262,7 @@ class TestRandomLayers:
         assert np.allclose(qnode1(weights), qnode2(weights), atol=tol)
 
     def test_random_layers_two_seeds_different(self, n_layers, tol):
-        """Test that pennylane.templates.layers.RandomLayers() does not have the same output for two different seeds."""
+        """Test that RandomLayers() does not have the same output for two different seeds."""
         n_rots = 10
         n_wires = 2
         dev = qml.device('default.qubit', wires=n_wires)
@@ -603,8 +300,8 @@ class TestRandomLayers:
         types = [type(q) for q in queue]
         assert len(types) - types.count(impr) == n_layers
 
-    def test_random_layer_imprimitive(self, ratio):
-        """Test that  pennylane.templates.layers.RandomLayer() has the right ratio of imprimitive gates."""
+    def test_random_layer_ratio_imprimitive(self, ratio):
+        """Test that  pennylane.templates.layers._random_layer() has the right ratio of imprimitive gates."""
         np.random.seed(12)
         n_rots = 500
         n_wires = 2
@@ -613,7 +310,8 @@ class TestRandomLayers:
         weights = np.random.randn(n_rots)
 
         def circuit(weights):
-            RandomLayer(weights=weights, wires=range(n_wires), ratio_imprim=ratio, impr=CNOT)
+            _random_layer(weights=weights, wires=range(n_wires), ratio_imprim=ratio,
+                          imprimitive=CNOT)
             return qml.expval(qml.PauliZ(0))
 
         qnode = qml.QNode(circuit, dev)
@@ -624,15 +322,15 @@ class TestRandomLayers:
         assert np.isclose(ratio_impr, ratio, atol=0.05)
 
     def test_random_layer_imprimitive(self, n_subsystems, impr, rots):
-        """Test that  pennylane.templates.layers.RandomLayer() uses the correct types of gates."""
+        """Test that  pennylane.templates.layers._random_layer() uses the correct types of gates."""
         np.random.seed(12)
         n_rots = 20
         dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         def circuit(weights):
-            RandomLayer(weights=weights, wires=range(n_subsystems),
-                                             imprimitive=impr, rotations=rots)
+            _random_layer(weights=weights, wires=range(n_subsystems),
+                          imprimitive=impr, rotations=rots)
             return qml.expval(qml.PauliZ(0))
 
         qnode = qml.QNode(circuit, dev)
@@ -644,7 +342,7 @@ class TestRandomLayers:
         assert unique == gates
 
     def test_random_layer_numgates(self, n_subsystems):
-        """Test that  pennylane.templates.layers.RandomLayer() uses the correct number of gates."""
+        """Test that  pennylane.templates.layers._random_layer() uses the correct number of gates."""
         np.random.seed(12)
         n_rots = 5
         impr = CNOT
@@ -652,7 +350,7 @@ class TestRandomLayers:
         weights = np.random.randn(n_rots)
 
         def circuit(weights):
-            RandomLayer(weights=weights, wires=range(n_subsystems), imprimitive=impr)
+            _random_layer(weights=weights, wires=range(n_subsystems), imprimitive=impr)
             return qml.expval(qml.PauliZ(0))
 
         qnode = qml.QNode(circuit, dev)
@@ -662,14 +360,14 @@ class TestRandomLayers:
         assert len(types) - types.count(impr) == n_rots
 
     def test_random_layer_randomwires(self, n_subsystems):
-        """Test that  pennylane.templates.layers.RandomLayer() picks random wires."""
+        """Test that  pennylane.templates.layers._random_layer() picks random wires."""
         np.random.seed(12)
         n_rots = 500
         dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         def circuit(weights):
-            RandomLayer(weights=weights, wires=range(n_subsystems))
+            _random_layer(weights=weights, wires=range(n_subsystems))
             return qml.expval(qml.PauliZ(0))
 
         qnode = qml.QNode(circuit, dev)
@@ -680,15 +378,15 @@ class TestRandomLayers:
         mean_wire = np.mean(wires_flat)
         assert np.isclose(mean_wire, (n_subsystems - 1) / 2, atol=0.05)
 
-    def test_random_layer_imprimitive(self, n_subsystems, tol):
-        """Test that  pennylane.templates.layers.RandomLayer() uses the correct weights."""
+    def test_random_layer_weights(self, n_subsystems, tol):
+        """Test that pennylane.templates.layers._random_layer() uses the correct weights."""
         np.random.seed(12)
         n_rots = 5
         dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         def circuit(weights):
-            RandomLayer(weights=weights, wires=range(n_subsystems))
+            _random_layer(weights=weights, wires=range(n_subsystems))
             return qml.expval(qml.PauliZ(0))
 
         qnode = qml.QNode(circuit, dev)
@@ -698,22 +396,3 @@ class TestRandomLayers:
         params_flat = [item for p in params for item in p]
         assert np.allclose(weights.flatten(), params_flat, atol=tol)
 
-    def test_random_layer_exception_subsystems(self):
-        """Tests that pennylane.templates.layers.RandomLayer() throws exception if n_wires < 2."""
-        np.random.seed(12)
-        n_rots = 2
-        n_wires = 1
-
-        dev = qml.device('default.qubit', wires=n_wires)
-        weights = np.random.randn(n_rots)
-
-        def circuit(weights):
-            RandomLayers(weights=weights, wires=range(n_wires))
-            return qml.expval(qml.PauliZ(0))
-
-        qnode = qml.QNode(circuit, dev)
-
-        with pytest.raises(ValueError) as excinfo:
-            qnode(weights)
-        assert excinfo.value.args[0] == "RandomLayer requires at least two wires or subsystems to apply " \
-                                        "the imprimitive gates."
